fix: HOST_NATIVE guards for m4aSoundInit, NULL guards, save/multiboot/librfu patches
- m4aSoundInit: delegate to HostNativeSoundInit on HOST_NATIVE to avoid
  VCount spin-waits, DMA register writes, and NUM_MUSIC_PLAYERS linker trick
- m4aSongNumStart/m4aSongNumStartOrChange: NULL-guard songHeader
- Various HOST_NATIVE compatibility patches for save, multiboot, librfu

+++ include/global.h
@@ -105,14 +105,16 @@
 // each of the files that were worked on.
 #define T1_READ_8(ptr)  ((ptr)[0])
 #define T1_READ_16(ptr) ((ptr)[0] | ((ptr)[1] << 8))
-#define T1_READ_32(ptr) ((ptr)[0] | ((ptr)[1] << 8) | ((ptr)[2] << 16) | ((ptr)[3] << 24))
-#define T1_READ_PTR(ptr) (u8 *) T1_READ_32(ptr)
+// Use unsigned casts to prevent sign-extension when casting the result to a pointer
+// on 64-bit native builds (important for HOST_NATIVE where pointers are 8 bytes).
+#define T1_READ_32(ptr) ((u32)(ptr)[0] | ((u32)(ptr)[1] << 8) | ((u32)(ptr)[2] << 16) | ((u32)(ptr)[3] << 24))
+#define T1_READ_PTR(ptr) ((u8 *)(uintptr_t)T1_READ_32(ptr))
 
 // T2_READ_8 is a duplicate to remain consistent with each group.
 #define T2_READ_8(ptr)  ((ptr)[0])
 #define T2_READ_16(ptr) ((ptr)[0] + ((ptr)[1] << 8))
-#define T2_READ_32(ptr) ((ptr)[0] + ((ptr)[1] << 8) + ((ptr)[2] << 16) + ((ptr)[3] << 24))
-#define T2_READ_PTR(ptr) (void *) T2_READ_32(ptr)
+#define T2_READ_32(ptr) ((u32)(ptr)[0] + ((u32)(ptr)[1] << 8) + ((u32)(ptr)[2] << 16) + ((u32)(ptr)[3] << 24))
+#define T2_READ_PTR(ptr) ((void *)(uintptr_t)T2_READ_32(ptr))
 
 // This macro is required to prevent the compiler from optimizing
 // a dpad up/down check in sub_812CAD8 (fame_checker.c).
